Update: Multi-threaded initialization of search array

+++ SearchChunk.cpp
@@ -1,12 +1,14 @@
+#include "SearchProblemSetup.cpp"
+
 void* SearchChunk(void* searchtask)
-{	
+{
 	SearchTask task = *((SearchTask*) searchtask); //convert search task pointer to search task object
 
 	int * Data = task.Data;
-	int Size = task.Size;
+	long Size = task.Size;
 	int SearchItem = task.SearchItem;
 
-	int count = 0; //how many times search item occurs in the data chunk
+	long count = 0; //how many times search item occurs in the data chunk
 
 	for(int i = 0; i < Size; i ++)
 	{
